+++ cypress/elements/layer.js
@@ -12,10 +12,18 @@ export class Layer {
         return this;
     }
 
-    selectTab(tab) {
-        cy.get('[data-test="dhis2-uicore-tabbar-tabs"]')
+    selectTab(tabName) {
+        cy.getByDataTest('dhis2-uicore-tabbar-tabs')
             .find('button')
-            .contains(tab)
+            .contains(tabName)
+            .click();
+
+        return this;
+    }
+
+    selectOu(ouName) {
+        cy.get('.tree-view.orgunit')
+            .contains(ouName)
             .click();
 
         return this;
@@ -41,21 +49,21 @@ export class Layer {
     }
 
     addToMap() {
-        cy.get('[data-test="dhis2-uicore-modalactions"]')
+        cy.getByDataTest('dhis2-uicore-modalactions')
             .contains('Add layer')
             .click();
     }
 
     validateDialogClosed(closed) {
         if (closed) {
-            cy.get('[data-test="layeredit"]').should('not.exist');
+            cy.getByDataTest('layeredit').should('not.exist');
         } else {
-            cy.get('[data-test="layeredit"]').should('be.visible');
+            cy.getByDataTest('layeredit').should('be.visible');
         }
     }
 
     validateCardTitle(title) {
-        cy.get('[data-test="layercard"]')
+        cy.getByDataTest('layercard')
             .contains(title)
             .should('be.visible');
 
@@ -63,7 +71,7 @@ export class Layer {
     }
 
     validateCardPeriod(period) {
-        cy.get('[data-test="layercard"]')
+        cy.getByDataTest('layercard')
             .contains(period)
             .should('be.visible');
 
@@ -72,7 +80,7 @@ export class Layer {
 
     validateCardItems(items) {
         items.forEach(item => {
-            cy.get('[data-test="layercard"]')
+            cy.getByDataTest('layercard')
                 .find('[data-test="layerlegend-item"]')
                 .contains(item)
                 .should('be.visible');