+++ src/components/map/MapName.js
@@ -4,10 +4,12 @@ import { connect } from 'react-redux';
 import i18n from '@dhis2/d2-i18n';
 import { IconClockHistory16 } from '@dhis2/ui';
 import { formatLocaleDate } from '../../util/time';
+import { useUserSettings } from '../UserSettingsProvider';
 import styles from './styles/MapName.module.css';
 
-const MapName = ({ showName, name, interpretationDate, uiLocale }) =>
-    showName && name ? (
+const MapName = ({ showName, name, interpretationDate }) => {
+    const { keyUiLocale } = useUserSettings();
+    return showName && name ? (
         <div className={styles.mapName}>
             <div className={`${styles.name} dhis2-maps-title`}>{name}</div>
             {interpretationDate && (
@@ -18,7 +20,7 @@ const MapName = ({ showName, name, interpretationDate, uiLocale }) =>
                         {
                             interpretationDate: formatLocaleDate(
                                 interpretationDate,
-                                uiLocale
+                                keyUiLocale || 'en'
                             ),
                         }
                     )}
@@ -26,17 +28,16 @@ const MapName = ({ showName, name, interpretationDate, uiLocale }) =>
             )}
         </div>
     ) : null;
-
-MapName.propTypes = {
-    showName: PropTypes.bool,
-    name: PropTypes.string,
-    interpretationDate: PropTypes.string,
-    uiLocale: PropTypes.string,
 };
 
-export default connect(({ map, download, settings }) => ({
+MapName.propTypes = {
+    name: PropTypes.string,
+    interpretationDate: PropTypes.string,
+    showName: PropTypes.bool,
+};
+
+export default connect(({ map, download }) => ({
     name: map.name,
     interpretationDate: map.interpretationDate,
     showName: download.showDialog ? download.showName : true,
-    uiLocale: settings.user.keyUiLocale,
 }))(MapName);